weekdays is not required, because KMA_REF With actual_size, give 5% wiggle room updated cnrm, hmcr, kma_ref, ncep, ukmo to check sizes.

+++ updatescripts/S2S/check_file_size.py
@@ -1,22 +1,29 @@
 import os
+import logging
 
-def process_file_by_size(filename, min_size, actual_size, dryrun=False, logging=None):
+def process_file_by_size(filename, min_size, actual_size, dryrun=False):
     # Check file size and remove if it doesn't match the expected size
     size = 0
+    # Give or take 5% of the actual size
+    actual_size_offset = int(actual_size*.05)
     if os.path.exists(filename):
         size = os.path.getsize(filename)
         # if actual_size == 0, we don't know the exact size, so don't check it.
         if actual_size is not None and 0 < actual_size != size:
-            if logging is not None:
-                logging.warning(f"target is wrong size ({size}), should be {actual_size}; removing {filename}")
-            if not dryrun:
-                os.unlink(filename)
-            return 0
+            if actual_size-actual_size_offset < size < actual_size+actual_size_offset:
+                return size
+            else:
+                logging.info(f"{filename} is wrong size ({size}), should be {actual_size}")
+                if not dryrun:
+                    logging.warning(f"removing wrong size file {filename}")
+                    os.unlink(filename)
+                return 0
         elif min_size is not None and size < min_size:
-            if logging is not None:
-                logging.warning(f"target too small ({size}), should be a minimum of {min_size}; removing {filename}")
+            logging.info(f"{filename} too small ({size}), should be a minimum of {min_size}")
             if not dryrun:
+                logging.warning(f"removing too small file {filename}")
                 os.unlink(filename)
             return 0
+    else:
+        logging.info(f"{filename} does not exist.")
     return size
-
